ai_services/pyproject.toml
[project]
name = "ai-services"
version = "0.1.0"
description = "RAG-based AI chatbot service for Selfish project management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "google-genai>=0.3.0",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "supabase>=2.11.1",
    "pillow>=11.0.0",
    "numpy>=2.2.0",
    "groq>=0.37.1",
]
